Arcade Mode and Challenge Mode
Split single game into two modes: arcade mode and challenge mode.
Challenge Mode: it is the classic mode which consists of multiple
levels. If the current level is clear, the player can enter the next
level.

Arcade Mode: it is a infinite game. Player is very unlikely to win in
this mode because the number of blocks will increase for every five
hits on the paddle. The player is only allowed to submit his score to
the server in this mode.